Refactorizacion ProductController
Sacamos los metodos de GetProduct a un controlador propio
